fix(sessions): unify legacy title projection
